requirements.txt
requests
python-dotenv
qrcode_terminal